impedance_controller node runs for the first time.

+++ ros2_ws/src/impedance_controller/include/impedance_controller_base.hpp
@@ -125,7 +125,7 @@ class ImpedanceControllerBase
 
         //! \name Actuator Parameters
         //! \{
-        double max_actuator_torque_;    //!< Maximal actuator torque (scalar) in Nm
+        int max_actuator_torque_;    //!< Maximal actuator torque (scalar) in Nm
         //! \}
 
         //! \name Input Variables
